clean up test directory
